requirements.txt
asgiref==3.8.1
Django==5.1.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1

django-crispy-forms
crispy-bootstrap5

python-dateutil
requests